Created ChatService
