+++ app/frontend/src/stores/ChatStore.tsx
@@ -1,34 +1,31 @@
 import { create } from 'zustand';
 import { produce } from 'immer'
 import { DEFAULT_CHAT_MODEL } from '../constants/models';
+import { buildDefaultChatHistory, buildDefaultModel } from './modelBuilders';
 
 type ChatStore = {
     currentChatIndex: number,
     currentChatHistory: ChatHistory,
+    chatHistoriesDescriptions: string[],
+    chatIndexToLoadOrDelete: number | null;
+    setChatIndexToLoadOrDelete: (index: number | null) => void;
     setCurrentChatHistory: (param: ChatHistory | ((prev: ChatHistory) => ChatHistory)) => void;
     setDefaultChatHistory: () => void;
     getDefaultModel: () => string;
     setCurrentChatIndex: (index: number) => void;
+    setChatHistoriesDescriptions: (descriptions: string[]) => void;
 };
-
-const buildDefaultModel = () => {
-    return DEFAULT_CHAT_MODEL
-}
-
-
-const buildDefaultChatHistory = () => {
-    const defaultChatHistory: ChatHistory = {
-        chatItems: [],
-        description: "",
-        uuid: "",
-        model: buildDefaultModel(),
-    };
-    return defaultChatHistory
-}
 
 export const useChatStore = create<ChatStore>((set) => ({
     currentChatIndex: 0,
+    chatIndexToLoadOrDelete: null,
     currentChatHistory: buildDefaultChatHistory(),
+    chatHistoriesDescriptions: ["Conversation 1"],
+    setChatIndexToLoadOrDelete: (index: number | null) => {
+        set((state) => produce(state, (draft) => {
+            draft.chatIndexToLoadOrDelete = index;
+        }));
+    },
     setCurrentChatHistory: (param: ChatHistory | ((prev: ChatHistory) => ChatHistory)) => {
         set((state) => produce(state, (draft) => {
             if (typeof param === "function") {
@@ -49,6 +46,11 @@ export const useChatStore = create<ChatStore>((set) => ({
     setCurrentChatIndex: (index: number) => {
         set((state) => produce(state, (draft) => {
             draft.currentChatIndex = index;
+        }));
+    },
+    setChatHistoriesDescriptions: (descriptions: string[]) => {
+        set((state) => produce(state, (draft) => {
+            draft.chatHistoriesDescriptions = descriptions;
         }));
     }
 }));
