+++ app/frontend/src/stores/modelBuilders.ts
@@ -1,0 +1,16 @@
+import { DEFAULT_CHAT_MODEL } from "../constants/models";
+
+export const buildDefaultModel = () => {
+    return DEFAULT_CHAT_MODEL
+}
+
+
+export const buildDefaultChatHistory = () => {
+    const defaultChatHistory: ChatHistory = {
+        chatItems: [],
+        description: "",
+        uuid: "",
+        model: buildDefaultModel(),
+    };
+    return defaultChatHistory
+}
